Changed: Merge batch, batch-base and register.
