+++ src/treemodel/delta/insertdelta.js
@@ -6,7 +6,7 @@
 'use strict';
 
 import Delta from './delta.js';
-import register from './register.js';
+import register from '../batch.js';
 import InsertOperation from '../operation/insertoperation.js';
 
 /**
